+++ attributes/default.rb
@@ -39,6 +39,8 @@ default['android-sdk']['license']['white_list']     = %w(android-sdk-license-.+ 
 default['android-sdk']['license']['black_list']     = []    # e.g. ['mips-.+', 'android-wear-sdk-license-.+']
 default['android-sdk']['license']['default_answer'] = 'n'   # 'y' or 'n' ('yes' or 'no')
 
-default['android-sdk']['script_addons']['path']     = '/usr/local/bin'
-default['android-sdk']['script_addons']['owner']    = node['android-sdk']['owner']
-default['android-sdk']['script_addons']['group']    = node['android-sdk']['group']
+default['android-sdk']['scripts']['path']          = '/usr/local/bin'
+default['android-sdk']['scripts']['owner']         = node['android-sdk']['owner']
+default['android-sdk']['scripts']['group']         = node['android-sdk']['group']
+
+default['android-sdk']['maven-rescue']             = false